styles(InputDay): migrate InputDate to vanilla css
