+++ apps/store/src/components/InputDay/InputDay.tsx
@@ -1,17 +1,24 @@
-import { keyframes } from '@emotion/react'
-import styled from '@emotion/styled'
 import * as Popover from '@radix-ui/react-popover'
+import { clsx } from 'clsx'
 import { sv } from 'date-fns/locale'
 import { useTranslation } from 'next-i18next'
-import { startTransition, useId, useState } from 'react'
+import { type CSSProperties, startTransition, useId, useState } from 'react'
 import { useInput, DayPicker, type SelectSingleEventHandler } from 'react-day-picker'
-import { ChevronIcon, theme, Text, LockIcon } from 'ui'
+import { ChevronIcon, Text, LockIcon, theme } from 'ui'
 import { LoadingDots } from '@/components/LoadingDots/LoadingDots'
 import { convertToDate } from '@/utils/date'
 import { Language } from '@/utils/l10n/types'
 import { useCurrentLocale } from '@/utils/l10n/useCurrentLocale'
 import { useFormatter } from '@/utils/useFormatter'
 import { useHighlightAnimation } from '@/utils/useHighlightAnimation'
+import {
+  trigger,
+  label,
+  innerWrapper,
+  chevronIcon,
+  popoverContent,
+  dayPicker,
+} from './InputDay.css'
 
 import 'react-day-picker/dist/style.css'
 
@@ -30,6 +37,7 @@ type Props = {
   required?: boolean
   autoFocus?: boolean
   loading?: boolean
+  className?: string
 }
 
 export const InputDay = (props: Props) => {
@@ -74,11 +82,16 @@ export const InputDay = (props: Props) => {
 
   return (
     <Popover.Root open={open} onOpenChange={handleOpenChange}>
-      <Wrapper {...animationProps} disabled={props.disabled} style={{ backgroundColor }}>
-        <Label htmlFor={inputId} data-disabled={props.disabled}>
+      <Popover.Trigger
+        {...animationProps}
+        className={clsx(trigger, props.className)}
+        style={{ backgroundColor }}
+        disabled={props.disabled}
+      >
+        <label className={label} htmlFor={inputId} data-disabled={props.disabled}>
           {props.label}
-        </Label>
-        <InnerWrapper>
+        </label>
+        <div className={innerWrapper}>
           {dateValue ? (
             <Text size="xl">{formatter.fromNow(dateValue)}</Text>
           ) : (
@@ -93,9 +106,9 @@ export const InputDay = (props: Props) => {
           ) : props.disabled ? (
             <LockIcon size="1rem" color={theme.colors.textSecondary} />
           ) : (
-            <StyledChevronIcon size="1rem" />
+            <ChevronIcon className={chevronIcon.animated} size="1rem" />
           )}
-        </InnerWrapper>
+        </div>
 
         <input
           {...inputProps}
@@ -104,146 +117,40 @@ export const InputDay = (props: Props) => {
           hidden={true}
           disabled={props.disabled}
         />
-      </Wrapper>
+      </Popover.Trigger>
 
       <Popover.Portal>
-        <PopoverContent side="bottom" align="start" sideOffset={-4} alignOffset={-4}>
+        <Popover.Content
+          className={popoverContent}
+          side="bottom"
+          align="start"
+          sideOffset={-4}
+          alignOffset={-4}
+        >
           {/* Fix issue on iOS where month dropdown is focused and can't be tapped */}
           <div tabIndex={1} aria-hidden={true} />
 
-          <StyledDayPicker
+          <DayPicker
             {...dayPickerProps}
+            className={dayPicker}
+            style={
+              {
+                '--rdp-accent-color': theme.colors.buttonPrimary,
+                '--rdp-background-color': theme.colors.gray200,
+                '--rdp-outline': `1px solid ${theme.colors.borderTranslucent1}`,
+              } as CSSProperties
+            }
             mode="single"
             captionLayout="dropdown-buttons"
             onSelect={handleSelect}
             weekStartsOn={1}
             components={{
-              IconRight: () => <ChevronRightIcon size="1rem" />,
-              IconLeft: () => <ChevronLeftIcon size="1rem" />,
+              IconRight: () => <ChevronIcon className={chevronIcon.right} size="1rem" />,
+              IconLeft: () => <ChevronIcon className={chevronIcon.left} size="1rem" />,
             }}
           />
-        </PopoverContent>
+        </Popover.Content>
       </Popover.Portal>
     </Popover.Root>
   )
 }
-
-const slideUpAndFadeAnimation = keyframes({
-  '0%': { opacity: 0, transform: 'translateY(10px)' },
-  '100%': { opacity: 1, transform: 'translateY(0)' },
-})
-
-const slideDownAndFadeAnimation = keyframes({
-  '0%': { opacity: 0, transform: 'translateY(-10px)' },
-  '100%': { opacity: 1, transform: 'translateY(0)' },
-})
-
-const PopoverContent = styled(Popover.Content)({
-  borderRadius: theme.radius.sm,
-  backgroundColor: theme.colors.backgroundStandard,
-  boxShadow: 'rgba(0, 0, 0, 0.06) 0px 2px 12px',
-
-  maxHeight: 'var(--radix-tooltip-content-available-height)',
-
-  animationDuration: '0.6s',
-  animationTimingFunction: 'cubic-bezier(0.16, 1, 0.3, 1)',
-  transformOrigin: 'var(--radix-tooltip-content-transform-origin)',
-
-  '&[data-side="top"]': {
-    animationName: slideUpAndFadeAnimation,
-  },
-
-  '&[data-side="bottom"]': {
-    animationName: slideDownAndFadeAnimation,
-  },
-})
-
-const StyledDayPicker = styled(DayPicker)({
-  // Docs: https://react-day-picker.js.org/basics/styling
-
-  '--rdp-accent-color': theme.colors.buttonPrimary,
-  '--rdp-background-color': theme.colors.gray200,
-  '--rdp-outline': `1px solid ${theme.colors.borderTranslucent1}`,
-
-  fontFamily: theme.fonts.body,
-  color: theme.colors.textPrimary,
-  margin: 0,
-  padding: theme.space.md,
-
-  '.rdp-day, .rdp-nav_button': {
-    height: 'calc(var(--rdp-cell-size) * 0.9)',
-    width: 'calc(var(--rdp-cell-size) * 0.9)',
-    borderRadius: theme.radius.xs,
-  },
-
-  '.rdp-day_today': {
-    fontWeight: 'normal',
-
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    backgroundColor: theme.colors.gray200,
-  },
-
-  '.rdp-caption_label': {
-    fontWeight: 'normal',
-    borderWidth: 1,
-  },
-
-  '.rdp-head_cell': {
-    color: theme.colors.textSecondary,
-    fontWeight: 'normal',
-  },
-})
-
-const ChevronRightIcon = styled(ChevronIcon)({ transform: 'rotate(-90deg)' })
-const ChevronLeftIcon = styled(ChevronIcon)({ transform: 'rotate(90deg)' })
-
-const Wrapper = styled(Popover.Trigger)({
-  position: 'relative',
-  borderRadius: theme.radius.sm,
-  height: '4.5rem',
-  width: '100%',
-  cursor: 'pointer',
-
-  ':focus-visible': {
-    boxShadow: theme.shadow.focus,
-    borderRadius: theme.radius.sm,
-  },
-})
-
-const InnerWrapper = styled.div({
-  width: '100%',
-  display: 'flex',
-  flexDirection: 'row',
-  alignItems: 'center',
-  justifyContent: 'space-between',
-  paddingTop: theme.space.xl,
-  paddingRight: '1.25rem',
-  paddingBottom: theme.space.sm,
-  paddingLeft: theme.space.md,
-})
-
-const Label = styled.label({
-  position: 'absolute',
-  left: theme.space.md,
-  top: theme.space.sm,
-  fontSize: theme.fontSizes.xs,
-  overflow: 'visible',
-  whiteSpace: 'nowrap',
-  textOverflow: 'ellipsis',
-  color: theme.colors.textSecondary,
-
-  '&&[data-disabled=true]': {
-    color: theme.colors.textSecondary,
-  },
-})
-
-const StyledChevronIcon = styled(ChevronIcon)({
-  pointerEvents: 'none',
-  transition: 'transform 200ms cubic-bezier(0.77,0,0.18,1)',
-
-  [`${Wrapper}[data-state=open] &`]: {
-    transform: 'rotate(180deg)',
-  },
-})
